Separate hpp and cpp files
